Implement AutoMapper for All Movie ViewModels and all Home Page and SingleMovie Page Entities

+++ src/Services/MovieDeck.Services.Data/IRatingsService.cs
@@ -12,6 +12,6 @@
 
         bool IsMovieRated(int movieId, string userId);
 
-        int GetUserRating(int movieId, string userID);
+        byte GetUserRating(int movieId, string userID);
     }
 }
